+++ packages/dialtone-vue3/components/rich_text_editor/rich_text_editor.stories.js
@@ -163,5 +163,6 @@ export const WithCustomExtensions = {
     allowCode: false,
     allowCodeblock: false,
     link: false,
+    useDivTags: false,
   },
 };