+++ packages/dialtone-vue3/recipes/buttons/callbar_button_with_popover/callbar_button_with_popover.vue
@@ -1,6 +1,7 @@
 <template>
   <div
     class="d-recipe-callbar-button-with-popover"
+    v-bind="addClassStyleAttrs($attrs)"
   >
     <dt-recipe-callbar-button
       :aria-label="ariaLabel"
@@ -37,7 +38,7 @@
       padding="none"
       class="d-recipe-callbar-button-with-popover__popover-wrapper"
       :dialog-class="['d-recipe-callbar-button-with-popover__popover', contentClass]"
-      v-bind="$attrs"
+      v-bind="removeClassStyleAttrs($attrs)"
       :open-popover="showPopover"
       @opened="onModalIsOpened"
     >
@@ -79,7 +80,7 @@ import { DtButton } from '@/components/button';
 import { DtPopover } from '@/components/popover';
 import { DtIconChevronUp } from '@dialpad/dialtone-icons/vue3';
 import { DtRecipeCallbarButton, CALLBAR_BUTTON_VALID_WIDTH_SIZE } from '../callbar_button';
-import utils, { warnIfUnmounted } from '@/common/utils';
+import utils, { warnIfUnmounted, removeClassStyleAttrs, addClassStyleAttrs, returnFirstEl } from '@/common/utils';
 
 export default {
   compatConfig: { MODE: 3 },
@@ -355,10 +356,12 @@ export default {
   },
 
   mounted () {
-    warnIfUnmounted(this.$el, this.$options.name);
+    warnIfUnmounted(returnFirstEl(this.$el), this.$options.name);
   },
 
   methods: {
+    removeClassStyleAttrs,
+    addClassStyleAttrs,
     arrowClick (ev) {
       this.$emit('arrow-click', ev);
       return this.toggleOpen();