+++ packages/dialtone-vue3/recipes/conversation_view/editor/editor_default.story.vue
@@ -15,6 +15,7 @@
       :cancel-set-link-button="$attrs.cancelSetLinkButton"
       :confirm-set-link-button="$attrs.confirmSetLinkButton"
       :remove-link-button="$attrs.removeLinkButton"
+      :use-div-tags="$attrs.useDivTags"
       :show-bold-button="$attrs.showBoldButton"
       :show-italics-button="$attrs.showItalicsButton"
       :show-strike-button="$attrs.showStrikeButton"
@@ -28,10 +29,12 @@
       :show-quote-button="$attrs.showQuoteButton"
       :show-quick-replies-button="$attrs.showQuickRepliesButton"
       :show-code-block-button="$attrs.showCodeBlockButton"
+      :show-inline-image-button="$attrs.showInlineImageButton"
       @focus="$attrs.onFocus"
       @blur="$attrs.onBlur"
       @input="$attrs.onInput"
       @quick-replies-click="$attrs.onQuickRepliesClick"
+      @inline-image-click="$attrs.onInlineImageClick"
     />
     <p><strong>Editor content is:</strong></p>
     <span>{{ value }}</span>