+++ packages/dialtone-emojis/CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.1.1](https://github.com/dialpad/dialtone/compare/dialtone-emojis/v1.1.0...dialtone-emojis/v1.1.1) (2025-04-02)
+
+
+### Bug Fixes
+
+* **Ci:** NO-JIRA changelog release order ([#689](https://github.com/dialpad/dialtone/issues/689)) ([bd4c080](https://github.com/dialpad/dialtone/commit/bd4c0802a59be6fdafb8f661a9c12c64ef764656))
+
 # [1.1.0](https://github.com/dialpad/dialtone/compare/dialtone-emojis/v1.0.8...dialtone-emojis/v1.1.0) (2024-12-13)